requirements.txt
drf-spectacular==0.27.1
djangorestframework-simplejwt[blacklist]==5.3.1 